+++ src/app/mypage/history/page.tsx
@@ -1,0 +1,13 @@
+import ResultList from "@/app/(features)/calculator/components/ResultList";
+
+import React from "react";
+
+const page = () => {
+  return (
+    <div>
+      <ResultList type="mypage" />
+    </div>
+  );
+};
+
+export default page;
